Fixing Up Layout
Added in a download resume button. Added  in a repo button for entire website (not finished). Removed an HTML file for simplicity to combine education and experience in one section. Removed the links for repo's for games. Removed information of Game Jam #1 and added in Final Project 2D instead. Updated the information for Final Project 3D. Added in two buttons to the home page instead of "Full Story"